Single module for Droplet power actions
