+++ tests/integration/targets/droplet_action_power/defaults/main.yml
