+++ tests/integration/targets/droplet_action_power/tasks/01_power_off.yml
@@ -1,0 +1,32 @@
+# code: language=ansible
+- name: 01_power_off | Testing block
+  block:
+    - name: 01_power_off | Power off Droplet
+      when: true
+      digitalocean.cloud.droplet_action_power:
+        token: "{{ digitalocean_token }}"
+        state: power_off
+        name: "{{ droplet_name }}"
+        region: "{{ droplet_region }}"
+      register: result
+
+    - name: 01_power_off | Verify power off Droplet
+      ansible.builtin.assert:
+        that:
+          - result.changed
+          - result.msg == "Droplet " ~ droplet_name ~ " (" ~ droplet.droplet.id ~ ") in " ~ droplet_region ~ " sent action 'power_off'"
+
+    - name: 01_power_off | Power off Droplet again
+      when: true
+      digitalocean.cloud.droplet_action_power:
+        token: "{{ digitalocean_token }}"
+        state: power_off
+        name: "{{ droplet_name }}"
+        region: "{{ droplet_region }}"
+      register: result
+
+    - name: 01_power_off | Verify power off Droplet idemopotency
+      ansible.builtin.assert:
+        that:
+          - not result.changed
+          - result.msg == "Droplet " ~ droplet_name ~ " (" ~ droplet.droplet.id ~ ") in " ~ droplet_region ~ " not sent action 'power_off', it is 'off'"
