+++ tests/integration/targets/droplet_action_power/tasks/02_power_on.yml
@@ -1,0 +1,32 @@
+# code: language=ansible
+- name: 02_power_on | Testing block
+  block:
+    - name: 02_power_on | Power on Droplet
+      when: true
+      digitalocean.cloud.droplet_action_power:
+        token: "{{ digitalocean_token }}"
+        state: power_on
+        name: "{{ droplet_name }}"
+        region: "{{ droplet_region }}"
+      register: result
+
+    - name: 02_power_on | Verify power on Droplet
+      ansible.builtin.assert:
+        that:
+          - result.changed
+          - result.msg == "Droplet " ~ droplet_name ~ " (" ~ droplet.droplet.id ~ ") in " ~ droplet_region ~ " sent action 'power_on'"
+
+    - name: 02_power_on | Power on Droplet again
+      when: true
+      digitalocean.cloud.droplet_action_power:
+        token: "{{ digitalocean_token }}"
+        state: power_on
+        name: "{{ droplet_name }}"
+        region: "{{ droplet_region }}"
+      register: result
+
+    - name: 02_power_on | Verify power on Droplet idemopotency
+      ansible.builtin.assert:
+        that:
+          - not result.changed
+          - result.msg == "Droplet " ~ droplet_name ~ " (" ~ droplet.droplet.id ~ ") in " ~ droplet_region ~ " not sent action 'power_on', it is 'active'"
