+++ tests/integration/targets/droplet_action_power/tasks/99_teardown.yml
