--- tests/integration/targets/droplet_action_power_off/tasks/01_power_off.yml
@@ -1,17 +1,0 @@
-# code: language=ansible
-- name: 01_power_off | Testing block
-  block:
-    - name: 01_power_off | Power off Droplet
-      when: true
-      digitalocean.cloud.droplet_action_power_off:
-        token: "{{ digitalocean_token }}"
-        state: present
-        name: "{{ droplet_name }}"
-        region: "{{ droplet_region }}"
-      register: result
-
-    - name: 01_power_off | Verify power off Droplet
-      ansible.builtin.assert:
-        that:
-          - result.changed
-          - result.msg == "Droplet " ~ droplet_name ~ " (" ~ result.action.resource_id ~ ") in " ~ droplet_region ~ " sent action 'power_off'"
