--- tests/integration/targets/droplet_action_power_on/tasks/01_power_on.yml
@@ -1,32 +1,0 @@
-# code: language=ansible
-- name: 01_power_on | Testing block
-  block:
-    - name: 01_power_on | Power off Droplet
-      when: true
-      digitalocean.cloud.droplet_action_power_off:
-        token: "{{ digitalocean_token }}"
-        state: present
-        name: "{{ droplet_name }}"
-        region: "{{ droplet_region }}"
-      register: result
-
-    - name: 01_power_on | Verify power off Droplet
-      ansible.builtin.assert:
-        that:
-          - result.changed
-          - result.msg == "Droplet " ~ droplet_name ~ " (" ~ result.action.resource_id ~ ") in " ~ droplet_region ~ " sent action 'power_off'"
-
-    - name: 01_power_on | Power on Droplet
-      when: true
-      digitalocean.cloud.droplet_action_power_on:
-        token: "{{ digitalocean_token }}"
-        state: present
-        name: "{{ droplet_name }}"
-        region: "{{ droplet_region }}"
-      register: result
-
-    - name: 01_power_on | Verify power off Droplet
-      ansible.builtin.assert:
-        that:
-          - result.changed
-          - result.msg == "Droplet " ~ droplet_name ~ " (" ~ result.action.resource_id ~ ") in " ~ droplet_region ~ " sent action 'power_on'"
