--- tests/integration/targets/droplet_action_shutdown/tasks/00_setup.yml
@@ -1,12 +1,0 @@
-# code: language=ansible
-- name: 00_setup | Testing block
-  block:
-    - name: 00_setup | Create Droplet
-      digitalocean.cloud.droplet:
-        token: "{{ digitalocean_token }}"
-        state: present
-        name: "{{ droplet_name }}"
-        region: "{{ droplet_region }}"
-        size: "{{ droplet_size }}"
-        image: "{{ droplet_image }}"
-        unique_name: true
