--- tests/integration/targets/droplet_action_shutdown/tasks/99_teardown.yml
@@ -1,10 +1,0 @@
-# code: language=ansible
-- name: 99_teardown | Testing block
-  block:
-    - name: 99_teardown | Delete Droplet
-      digitalocean.cloud.droplet:
-        state: absent
-        token: "{{ digitalocean_token }}"
-        name: "{{ droplet_name }}"
-        region: "{{ droplet_region }}"
-        unique_name: true
